# break out tuple wrapper to speed up centroid calc, 100x speedup