@tlnagy 2019-05-22
@johnnychen94 I just saw that you were planning to move this code over to ImageMorphology. Should I rebase this once you merge? Or can you rebase your PR?