@tlnagy 2019-05-22
This seemed a little weird to me and I found https://github.com/JuliaImages/Images.jl/pull/678 and https://github.com/JuliaImages/Images.jl/issues/677, which both in a single digit ms range. This makes me think that this is potentially a regression in base? @Evizero any thoughts?

My environment:

```julia
julia> versioninfo()
Julia Version 1.1.0
Commit 80516ca202 (2019-01-21 21:24 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: AMD Ryzen 7 2700X Eight-Core Processor
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, znver1)

```