@johnnychen94 2019-05-23
> Should I rebase this once you merge? Or can you rebase your PR?

I guess you can reopen this PR in ImageMorphology. Related codes (except test cases) will be removed in next minor release.