@johnnychen94 2019-05-29
Perhaps this is related to https://github.com/JuliaLang/julia/issues/28980

Can you reopen this PR in ImageMorphology? We can make new patches there since these functions are already ported and will be removed in next minor release of Images.jl

Thanks.