@tlnagy 2019-06-05
https://github.com/JuliaLang/julia/issues/28980 seems unrelated because the 100x fold regression is fixed on master but this PR still yields about a 10% increase in median speed:

### Images master
```julia
julia> a = rand(0:250, 1024, 1024);

julia> @benchmark Images.component_centroids(a)
BenchmarkTools.Trial: 
  memory estimate:  10.17 KiB
  allocs estimate:  4
  --------------
  minimum time:     1.972 ms (0.00% GC)
  median time:      2.748 ms (0.00% GC)
  mean time:        2.742 ms (0.00% GC)
  maximum time:     5.276 ms (0.00% GC)
  --------------
  samples:          1819
  evals/sample:     1
```

### This PR

```julia
julia> @benchmark Images.component_centroids(a)
BenchmarkTools.Trial: 
  memory estimate:  10.31 KiB
  allocs estimate:  5
  --------------
  minimum time:     1.907 ms (0.00% GC)
  median time:      2.498 ms (0.00% GC)
  mean time:        2.489 ms (0.00% GC)
  maximum time:     3.595 ms (0.00% GC)
  --------------
  samples:          2000
  evals/sample:     1

julia> println(VERSION)
1.3.0-DEV.320
```