@tlnagy 2019-06-05
Moved to https://github.com/JuliaImages/ImageMorphology.jl/pull/20